Fix onError() throwing if err is undefined (#5)
While it's not very nice thing to do, it's totally allowed by
ECMA-262 to call `Promise.reject()` with an argument that is not
an instance of `Error`. If this happens inside a promise that
`async-retry-ng` is awaiting for then `onRetry()` might throw
depending on the exact type of the `reason` argument. Therefore
`onRetry()` should first check that `reason` is of expected type.

Fixes #4